rag application with langchain/requirements.txt
langchain
langchain_community
langchain_openai
openai
python-dotenv
chromadb
tiktoken
google-generativeai
langchain-google-genai
pypdf>=5.4.0
